+++ src/Pororoca.Domain/Features/Entities/Pororoca/PororocaCollectionItem.cs
@@ -4,17 +4,12 @@ namespace Pororoca.Domain.Features.Entities.Pororoca;
 
 public abstract class PororocaCollectionItem : ICloneable
 {
-    [JsonPropertyOrder(-2)]
-    [JsonInclude]
-    public Guid Id { get; init; }
-
     [JsonPropertyOrder(-1)]
     [JsonInclude]
     public string Name { get; protected set; }
 
-    protected PororocaCollectionItem(Guid id, string name)
+    protected PororocaCollectionItem(string name)
     {
-        Id = id;
         Name = name;
     }